Conflicts:
	features/cerberusweb.timetracking/plugin.xml

+++ features/cerberusweb.core/api/uri/tickets.php
@@ -668,6 +668,9 @@ class ChTicketsPage extends CerberusPageExtension {
 		if(!empty($to))
 			$params['to'] = $to;
 			
+		if(empty($subject) && empty($content))
+			return json_encode(array());
+			
 		@$type = DevblocksPlatform::importGPC($_REQUEST['type'],'string','');
 		
 		switch($type) {
@@ -745,6 +748,7 @@ class ChTicketsPage extends CerberusPageExtension {
 		@$active_worker = CerberusApplication::getActiveWorker();
 		
 		if(!empty($draft_id)
+			&& null != ($draft = DAO_MailQueue::get($draft_id))
 			&& ($active_worker->id == $draft->worker_id || $active_worker->is_superuser)) {
 			
 			DAO_MailQueue::delete($draft_id);
@@ -1173,6 +1177,7 @@ class ChTicketsPage extends CerberusPageExtension {
         $query = trim($query);
         
         $visit = CerberusApplication::getVisit(); /* @var $visit CerberusVisit */
+		$active_worker = CerberusApplication::getActiveWorker();
 		$searchView = C4_AbstractViewLoader::getView(CerberusApplication::VIEW_SEARCH);
 		
 		if(null == $searchView)
@@ -1226,6 +1231,10 @@ class ChTicketsPage extends CerberusPageExtension {
                 break;
                 
         }
+        
+		// Force group ACL
+		if(!$active_worker->is_superuser)
+        	$params[SearchFields_Ticket::TICKET_TEAM_ID] = new DevblocksSearchCriteria(SearchFields_Ticket::TICKET_TEAM_ID, 'in', array_keys($active_worker->getMemberships()));
         
         $searchView->params = $params;
         $searchView->renderPage = 0;
